+++ src/domains/lightning/usecases/service/mod.rs
@@ -4,7 +4,7 @@ mod lightning_node;
 mod lightning_payments;
 mod lightning_payments_processor;
 mod lightning_service;
-mod wallet;
+mod lightning_wallet;
 
 pub use lightning_payments_processor::LightningPaymentsProcessor;
 pub use lightning_service::LightningService;
